Updating publish instance metrics to append new configurations to existing logging configurations
This change is to fix issue where customers are not seeing logs after running LCC. Previously with
AL1, there were two cloudwatch instances running where the old process had old logging configs and
was publishing logs, while the newer process had the metrics configs and was publishing metrics.
For AL2 instead of starting a second process, it is actually restarting the old process, which results
in that new process only having metrics configs and not having any logging configs. As a result no
logs are emitted. This change will append the new configs so the process can publish both the logs
and the metrics.